added middleware and changed interceptors
