+++ src/app/auth/Auth.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useState } from 'react'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import LOGO from "../../images/logo.svg"
