+++ src/app/car/add-car/page.tsx
@@ -1,8 +1,16 @@
-import Cars from "./Cars";
+import FormCar from "./FormCar";
+import TableCar from "./TableCar";
+import styles from "./Cars.module.css";
+
 export default function CarsPage() {
   return (
-    <div>
-      <Cars />
+    <div className={styles.items}>
+      <div className={styles.main_form}>
+        <div className={styles.container}>
+          <FormCar />
+          <TableCar />
+        </div>
+      </div>
     </div>
   );
 }
