+++ src/app/layout.tsx
@@ -1,10 +1,9 @@
 import type { Metadata } from "next";
-import { Inter } from 'next/font/google'
+import { Montserrat } from 'next/font/google'
 import "./globals.css";
-import { QueryClient } from "@tanstack/react-query";
 import { Providers } from "./providers";
 
-const zen = Inter({
+const zen = Montserrat({
 	subsets: ['cyrillic', 'latin'],
 	weight: ['300', '400', '500', '600', '700'],
 	display: 'swap',
@@ -22,16 +21,12 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const queryClient = new QueryClient();
 
   return (
     <html lang="en">
       <body
         className={`${zen.variable} antialiased`}
       >
-        {/* <QueryClientProvider client={queryClient}>
-					{children}
-				</QueryClientProvider> */}
         <Providers>
           {children}
         </Providers>
