+++ src/services/auth-token.service.ts
@@ -1,25 +1,41 @@
+import Cookies from 'js-cookie'
+
 export enum EnumTokens {
 	'ACCESS_TOKEN' = 'accessToken',
 	'REFRESH_TOKEN' = 'refreshToken'
 }
 
 export const getAccessToken = () => {
-	const accessToken = localStorage.getItem(EnumTokens.ACCESS_TOKEN)
-	console.log(accessToken);
+	const accessToken = Cookies.get(EnumTokens.ACCESS_TOKEN)
+	// const accessToken = localStorage.getItem(EnumTokens.ACCESS_TOKEN)
 	return accessToken || null
 }
 
 export const getRefreshToken = () => {
-	const refreshToken = localStorage.getItem(EnumTokens.REFRESH_TOKEN)
+	const refreshToken = Cookies.get(EnumTokens.REFRESH_TOKEN)
+	// const refreshToken = localStorage.getItem(EnumTokens.REFRESH_TOKEN)
 	return refreshToken || null
 }
 
 export const saveTokenStorage = (accessToken: string, refreshToken: string) => {
-	localStorage.setItem(EnumTokens.ACCESS_TOKEN, accessToken)
-	localStorage.setItem(EnumTokens.REFRESH_TOKEN, refreshToken)
+	Cookies.set(EnumTokens.ACCESS_TOKEN, accessToken, {
+		domain: 'localhost',
+		expires: 1
+	})
+
+	Cookies.set(EnumTokens.REFRESH_TOKEN, refreshToken, {
+		domain: 'localhost',
+		expires: 1
+	})
+
+	// localStorage.setItem(EnumTokens.ACCESS_TOKEN, accessToken)
+	// localStorage.setItem(EnumTokens.REFRESH_TOKEN, refreshToken)
 }
 
 export const removeFromStorage = () => {
-	localStorage.removeItem(EnumTokens.ACCESS_TOKEN)
-	localStorage.removeItem(EnumTokens.REFRESH_TOKEN)
+	Cookies.remove(EnumTokens.ACCESS_TOKEN)
+	Cookies.remove(EnumTokens.REFRESH_TOKEN)
+
+	// localStorage.removeItem(EnumTokens.ACCESS_TOKEN)
+	// localStorage.removeItem(EnumTokens.REFRESH_TOKEN)
 }
