+++ src/services/auth.service.ts
@@ -26,22 +26,12 @@ export const authService = {
 	async checkAuth() {
 		const refreshToken = getRefreshToken()
 
-		const response = await axiosWithAuth.post<IAuthRefreshResponse>(`/auth/refresh`, { refreshToken })
+		const response = await axiosWithAuth.post<IAuthRefreshResponse>(`/auth/refresh`, { refresh_token: refreshToken })
 
 		if (response.data.access_token && response.data.refresh_token) {
 			saveTokenStorage(response.data.access_token, response.data.refresh_token)
 		}
-	}
-
-	// async getNewTokens() {
-	// 	const response = await axiosWithAuth.post<IAuthLoginResponse>(
-	// 		'/auth/login/access-token' // auth/refresh
-	// 	)
-
-	// 	if (response.data.accessToken) saveTokenStorage(response.data.accessToken)
-
-	// 	return response
-	// },
+	},
 
 	// async logout() {
 	// 	const response = await axiosWithAuth.post<boolean>('/auth/logout')
