+++ src/types/car.type.ts
@@ -1,0 +1,10 @@
+export interface IСhoiceCar {
+  stamp: string;
+  model: string;
+  year: string;
+  number: string;
+  kpp: string;
+  classCar: string;
+  price: string;
+  text: string;
+}
